Indicate active env name in spawned environment
So that subprocesses (especially interactive shells) can be aware they
are within a Vaulted-spawned environment, set an environment variable in
spawned environments containing the name of the vaulted environment.

By default the environment variable is VAULTED_ACTIVE_ENV, but this is
configurable at the command line via the -E flag. Specifying an empty
string instead (eg, -E '') will prevent the setting of such a variable
in spawned environments.