Fix suggested fish command to load variables
The previous command ended up running all of the lines together on a
single line and since the first line begins with a comment character,
the entire script was effectively commented out!

For additional information, see:
 * https://github.com/fish-shell/fish-shell/issues/159
 * https://github.com/fish-shell/fish-shell/issues/3993

Additionally, commands run in a fish command substitution do not have
access to the TTY, preventing vaulted from prompting for a password.
Using piping avoids this issue all together.

For additional information, see:
 * https://github.com/fish-shell/fish-shell/issues/1362
 * https://github.com/fish-shell/fish-shell/pull/3922