Move role assumption out of session generation
It was very intentional that role assumptions be done _after_ the
session credentials were generated and cached. This was to allow the
session credentials to be reused when assuming a role again. However,
the role assumption has still been being done by the session generation
code, limiting flexibility with how the role assumption is done.

To alleviate some of these restrictions, the session creation no longer
assumes the role configured in the vault and instead returns the role
in the session. It is now left to the caller to ensure the role is
assumed after the session is generated.

Allowing a mutated vault to be passed into the session generation was
also considered. However, this would have opened up the session cache to
integrity violations, resulting in very difficult to track down issues.