Invert session generation to require a vault
This inversion provides an opportunity to modify the vault after loading
it from disk, but before generating a session. This must be used with
caution as the generated session could be vastly different than the
contents of the vault itself.

It is intended that session caching be extended to allow caching of
multiple sessions for a single vault. Each cached session will be keyed
on the contents of the vault used to generate the session.