add vault key signing to ssh agent
this adds the ability to add the vault signing url and valid
principals to a vaulted vault to enable hashicorp vault to sign
ssh keys for a set of principals.

this does not make the signing url or the valid principals
available in the edit menu or command flags, which will come in a
later commit.