handle caching of generated SSH key
to maintain a quality user experience we should cache the
in-memory generated SSH key to avoid re-creating it every time
we open the vault.  this is both a more intuitive experience
and a time save.

now the generated key will persist as long as a vault sessions
does and recreate only as needed.